backend/requirements.txt
# ── Core Framework ─────────────────────────────────────────────────────────────
fastapi==0.110.2
uvicorn[standard]==0.29.0
gunicorn==22.0.0
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Database (Anshul's domain — listed here for shared requirements.txt) ───────
sqlalchemy==2.0.29
alembic==1.13.1
asyncpg==0.29.0

# ── Redis ──────────────────────────────────────────────────────────────────────
redis==5.0.4

# ── Celery ─────────────────────────────────────────────────────────────────────
celery==5.4.0

# ── AI / Agent Framework ───────────────────────────────────────────────────────
langchain==0.2.1
langchain-openai==0.1.7
langgraph==0.1.5
openai==1.30.1
qdrant-client==1.9.1
tiktoken==0.7.0

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0
json-repair==0.25.2       # Fallback for malformed LLM JSON output

# ── Testing ─────────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0             # Used in FastAPI test client

beautifulsoup4==4.12.3
lxml==4.9.3
duckduckgo-search==5.3.1
python-socketio==5.11.2
